feat(mood): renamed moodform step files, constants, commented out useMoods' useEffect
