+++ src/constants/moodConstants.ts
